Retrieve git build at compilation time
