+++ src/cache.rs
@@ -68,6 +68,8 @@ pub async fn fill(
     map.insert("SUCCESS_EMOJI_NAME", env::var("SUCCESS_EMOJI_NAME")?);
     map.insert("LOADING_EMOJI_ID", env::var("LOADING_EMOJI_ID")?);
     map.insert("LOADING_EMOJI_NAME", env::var("LOADING_EMOJI_NAME")?);
+    map.insert("GIT_HASH_LONG", String::from(env!("GIT_HASH_LONG")));
+    map.insert("GIT_HASH_SHORT", String::from(env!("GIT_HASH_SHORT")));
     map.insert("JOIN_LOG", env::var("JOIN_LOG")?);
     map.insert("BOT_PREFIX", String::from(prefix));
     map.insert("BOT_ID", id.to_string());
