+++ src/commands/botinfo.rs
@@ -4,7 +4,6 @@ use serenity::model::prelude::*;
 use serenity::prelude::*;
 
 use std::env;
-use std::process::Command;
 
 use crate::cache::ConfigCache;
 use crate::utls::constants::COLOR_OKAY;
@@ -15,6 +14,8 @@ pub async fn botinfo(ctx: &Context, msg: &Message, _args: Args) -> CommandResult
     let topgg = env::var("DISCORDBOTS_LINK").expect("Expected top.gg link envvar");
     let github = env::var("GITHUB_LINK").expect("Expected github link envvar");
     let stats = env::var("STATS_LINK").expect("Expected stats link envvar");
+    let hash_short = env::var("GIT_HASH_SHORT").expect("Expected stats link envvar");
+    let hash_long = env::var("GIT_HASH_LONG").expect("Expected stats link envvar");
 
     let avatar = {
         let data_read = ctx.data.read().await;
@@ -53,12 +54,9 @@ pub async fn botinfo(ctx: &Context, msg: &Message, _args: Args) -> CommandResult
                 e.thumbnail(avatar);
                 e.color(COLOR_OKAY);
 
-                let hash = get_github_build(false);
-                let short = get_github_build(true);
-                let str = format!(
-                    "Built from commit [{}]({}{}{})",
-                    short, github, "/commit/", hash
-                );
+                let str = format!("Built from commit [{}]({}{}{})",
+                                  hash_short, github, "/commit/", hash_long);
+
                 e.fields(vec![
                     ("Language", "Rust 2018", false),
                     ("Software Version", env!("CARGO_PKG_VERSION"), false),
@@ -77,16 +75,4 @@ pub async fn botinfo(ctx: &Context, msg: &Message, _args: Args) -> CommandResult
 
     debug!("Command executed");
     Ok(())
-}
-
-pub fn get_github_build(short: bool) -> String {
-    // note: add error checking yourself.
-    let mut args = vec!["rev-parse"];
-    if short {
-        args.push("--short");
-    }
-
-    args.push("HEAD");
-    let output = Command::new("git").args(&args).output().unwrap();
-    String::from_utf8(output.stdout).unwrap()
 }
